requirements.txt
streamlit
groq
langchain
langchain-community
chromadb
pypdf
sentence-transformers
python-dotenv